feat: cli remove validate, absorb batch into run --batch
Rimosso `toolkit validate` (coperto da run --dry-run).
Assorbito `toolkit batch` in `toolkit run --batch <file>`.
Aggiunto smoke/batch.txt con smoke test offline.
Backward compat: batch come comando hidden + deprecation warning.

- toolkit/cli/cmd_validate.py eliminato
- toolkit/cli/cmd_run.py: _run_batch(), flag --batch, helper condivisi
- toolkit/cli/cmd_batch.py: thin wrapper deprecato (272 -> 65 righe)
- toolkit/cli/app.py: remove validate registration
- smoke/batch.txt: batch list per smoke offline
- Test: aggiornati per validate rimosso e run --batch
- Nuovo test: test_run_batch_end_to_end (@pytest.mark.contract)

Bilancio: -32 netto righe, 116 test passano, ruff OK
