+++ toolkit/cli/app.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import typer
 
 from toolkit.cli.cmd_run import register as register_run
-from toolkit.cli.cmd_validate import register as register_validate
 from toolkit.cli.cmd_inspect import register as register_inspect
 from toolkit.cli.cmd_batch import register as register_batch
 from toolkit.cli.cmd_contract import register as register_contract
@@ -25,7 +24,6 @@ def _main_options(
 
 # registra comandi
 register_run(app)
-register_validate(app)
 register_inspect(app)
 register_batch(app)
 register_contract(app)
